order of off canvas elements
I may have been being a bit of an idiot here, but I just lost over an hour trying to debug a problem with my markup before finally realising that the order of the elements is important. I usually place a nav after my main content if possible, it may be myth but apparently it's good for search bots to get to your content asap and a fixed position nav could go anywhere in the dom. 

Unfortunately the default push animation was nowhere to be found and I tried absolutely everything before I got around to seeing if the order of the elements made a difference!